requirements.txt
streamlit
python-dotenv
gtts
google-generativeai
python-dotenv